Final changes after move.
